T293: cleaning up parseBytes
