+++ src/main/scala/scorex/transaction/PaymentTransaction.scala
@@ -108,17 +108,8 @@ object PaymentTransaction {
     }
   }
 
-  def parseBytes(data: Array[Byte]): Try[PaymentTransaction] = {
-    data.head match {
-      case transactionType: Byte if transactionType == TransactionType.PaymentTransaction.id =>
-        parseTail(data.tail)
-      case transactionType =>
-        Failure(new Exception(s"Incorrect transaction type '$transactionType' in PaymentTransaction data"))
-    }
-  }
 
-  def parseTail(data: Array[Byte]): Try[PaymentTransaction] =
-    Try {
+  def parseTail(data: Array[Byte]): Try[PaymentTransaction] =    Try {
       require(data.length >= BaseLength, "Data does not match base length")
 
       var position = 0
