+++ src/test/scala/scorex/transaction/GenesisTransactionSpecification.scala
@@ -5,10 +5,23 @@ import org.scalatest._
 import org.scalatest.prop.PropertyChecks
 import scorex.account.{PrivateKeyAccount, PublicKeyAccount}
 import scorex.crypto.encode.Base58
+import scorex.transaction.TransactionParser.TransactionType
+
+import scala.util.{Failure, Try}
 
 class GenesisTransactionSpecification extends PropSpec with PropertyChecks with Matchers {
 
   private val defaultRecipient = new PublicKeyAccount(Array.fill(32)(0: Byte))
+
+
+  def parseBytes(data: Array[Byte]): Try[GenesisTransaction] = {
+    data.head match {
+      case transactionType: Byte if transactionType == TransactionType.GenesisTransaction.id =>
+        GenesisTransaction.parseTail(data.tail)
+      case transactionType =>
+        Failure(new Exception(s"Incorrect transaction type '$transactionType' in GenesisTransaction data"))
+    }
+  }
 
   property("GenesisTransaction Signature should be the same") {
     val balance = 457L
@@ -21,10 +34,11 @@ class GenesisTransactionSpecification extends PropSpec with PropertyChecks with 
     assert(actual == expected)
   }
 
+
   property("GenesisTransaction parse from Bytes should work fine") {
     val bytes = Base58.decode("5GoidXWjBfzuS9tZm4Fp6GAXUYFunVMsoWAew3eBnDbmaDi7WiP9yVpBD2").get
 
-    val actualTransaction = GenesisTransaction.parseBytes(bytes).get
+    val actualTransaction = parseBytes(bytes).get
 
     val balance = 12345L
     val timestamp = 1234567890L
@@ -39,7 +53,7 @@ class GenesisTransactionSpecification extends PropSpec with PropertyChecks with 
         val recipient = new PrivateKeyAccount(recipientSeed)
         val source = GenesisTransaction.create(recipient, amount, time).right.get
         val bytes = source.bytes
-        val dest = GenesisTransaction.parseBytes(bytes).get
+        val dest = parseBytes(bytes).get
 
         source should equal(dest)
     }
