+++ src/test/scala/scorex/transaction/LeaseTransactionSpecification.scala
@@ -2,13 +2,22 @@ package scorex.transaction
 
 import org.scalatest._
 import org.scalatest.prop.PropertyChecks
+import scorex.transaction.TransactionParser.TransactionType
 import scorex.transaction.lease.LeaseTransaction
+
+import scala.util.Try
 
 class LeaseTransactionSpecification extends PropSpec with PropertyChecks with Matchers with TransactionGen {
 
+  def parseBytes(bytes: Array[Byte]): Try[LeaseTransaction] = Try {
+    require(bytes.head == TransactionType.LeaseTransaction.id)
+    LeaseTransaction.parseTail(bytes.tail).get
+  }
+
+
   property("Lease transaction serialization roundtrip") {
     forAll(leaseGenerator) { tx: LeaseTransaction =>
-      val recovered = LeaseTransaction.parseBytes(tx.bytes).get
+      val recovered = parseBytes(tx.bytes).get
 
       assertTxs(recovered, tx)
     }
